+++ apm-agent-plugins/apm-rabbitmq/apm-rabbitmq-spring/src/test/java/co/elastic/apm/agent/rabbitmq/RabbitMqTestBase.java
@@ -1,0 +1,64 @@
+/*-
+ * #%L
+ * Elastic APM Java agent
+ * %%
+ * Copyright (C) 2018 - 2021 Elastic and contributors
+ * %%
+ * Licensed to Elasticsearch B.V. under one or more contributor
+ * license agreements. See the NOTICE file distributed with
+ * this work for additional information regarding copyright
+ * ownership. Elasticsearch B.V. licenses this file to you under
+ * the Apache License, Version 2.0 (the "License"); you may
+ * not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ * #L%
+ */
+package co.elastic.apm.agent.rabbitmq;
+
+import co.elastic.apm.agent.AbstractInstrumentationTest;
+import co.elastic.apm.agent.bci.ElasticApmAgent;
+import org.junit.AfterClass;
+import org.springframework.amqp.rabbit.core.RabbitTemplate;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.util.TestPropertyValues;
+import org.springframework.context.ApplicationContextInitializer;
+import org.springframework.context.ConfigurableApplicationContext;
+import org.testcontainers.containers.RabbitMQContainer;
+
+public class RabbitMqTestBase extends AbstractInstrumentationTest {
+
+    protected static RabbitMQContainer container = new RabbitMQContainer(TestConstants.DOCKER_TESTCONTAINER_RABBITMQ_IMAGE);
+
+    static class Initializer implements ApplicationContextInitializer<ConfigurableApplicationContext> {
+        @Override
+        public void initialize(ConfigurableApplicationContext configurableApplicationContext) {
+            container.withExtraHost("localhost", "127.0.0.1");
+            container.start();
+
+            TestPropertyValues.of(
+                "spring.rabbitmq.host=" + container.getHost(),
+                "spring.rabbitmq.port=" + container.getAmqpPort(),
+                "spring.rabbitmq.username=" + container.getAdminUsername(),
+                "spring.rabbitmq.password=" + container.getAdminPassword())
+                .applyTo(configurableApplicationContext.getEnvironment());
+        }
+    }
+
+    @AfterClass
+    public static void after() {
+        container.close();
+        ElasticApmAgent.reset();
+    }
+
+    @Autowired
+    public RabbitTemplate rabbitTemplate;
+}
